+++ src/psx/input/dualanalog.cpp
@@ -121,7 +121,7 @@ void InputDevice_DualAnalog::StateAction(StateMem* sm, const unsigned load, cons
   Power();
  else if(load)
  {
-  if((transmit_pos + transmit_count) > sizeof(transmit_buffer))
+  if(((uint64)transmit_pos + transmit_count) > sizeof(transmit_buffer))
   {
    transmit_pos = 0;
    transmit_count = 0;
