+++ src/psx/input/dualshock.cpp
@@ -274,7 +274,7 @@ void InputDevice_DualShock::StateAction(StateMem* sm, const unsigned load, const
   Power();
  else if(load)
  {
-  if((transmit_pos + transmit_count) > sizeof(transmit_buffer))
+  if(((uint64)transmit_pos + transmit_count) > sizeof(transmit_buffer))
   {
    transmit_pos = 0;
    transmit_count = 0;
